app/requirements.txt
python-dotenv
elasticsearch
celery
hazm
